Create VirtualMemorySimulatorPart3
This is a code for virtual memory simulator in which size of page table is 256x256 bytes and size of physical memory is 128x256 bytes. This code also employes mechanism of Least Recently Used (LRU) and also writes back the least recently used page to the swap space. That is done by employing concept of dirty bit.